Include Travis configuration.
Rename dns-bridge.go -> bridge.go.

DNS-SD is redundant with the base name.

Refactor static state to configuration objects.

We should not use init for state that can be easily wrapped in an
object.  This limits testability and introduces side-effects if
somebody needs to link to this code---for whatever reason should
occur.